# REF/PERF: PeriodDtype decouple from DateOffset