@jbrockmendel 2020-06-01
> The other methods of the PeriodDtype class don't seem used and/or tested?

yah not yet.  i can remove them until they are actually used i guess