@jbrockmendel 2020-06-01
> It needs another name IMO, since we already have another PeriodDtype class that is something else

Any suggestions?  (I plan to mix it into the core.dtypes PeriodDtype before long)