@jbrockmendel 2020-06-02
updated with a placeholder name PeriodPseudoDtype