@jorisvandenbossche 2020-06-02
Alternative could also be `PeriodCode` 
(but if you are still going to change it, might also not matter too much)