@jbrockmendel 2020-06-02
updated with docstring.  im increasingly confident this will (eventually, soon-ish) let us de-duplicate Resolution, FreqGroup, and some cdef constants we define in libperiod.  Maybe even NPY_DATETIMEUNIT depending on what numpy does